Improve logic for Placeholders and Defaults
